+++ src/components/layout/LocationSearchPopup.tsx
@@ -209,6 +209,7 @@ export function LocationSearchPopup({
                         hit.countryCode === "DE" ||
                         hit.countryCode === "FR" ||
                         hit.countryCode === "DK" ||
+                        hit.countryCode === "CH" ||
                         hit.countryCode === "CZ" ||
                         hit.countryCode === "PL") && (
                         <span className="ml-1.5 text-[11px] font-semibold text-[#ff385c]">